requirements.txt
fastapi
python-pptx
beautifulsoup4
requests
openai
python-dotenv
aiohttp
supabase
python-multipart
uvicorn